pyproject.toml
[project]
name = "product-price-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "lxml>=6.0.0",
    "pip>=24.0",
    "playwright>=1.53.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0",
]
